@@ -15,14 +15,12 @@
 package com.example.android.basicandroidaccessibility;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.ImageButton;
+import android.widget.Toolbar;
 
-/**
- *
- */
+import androidx.appcompat.app.AppCompatActivity;
+
 public class ContentLabelingActivity extends AppCompatActivity {
 
     private boolean mPlaying = false;
@@ -32,10 +30,10 @@ public class ContentLabelingActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_content_labeling);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        setActionBar(toolbar);
 
-        mPlayPauseToggleImageView = (ImageButton) findViewById(R.id.play_pause_toggle_view);
+        mPlayPauseToggleImageView = findViewById(R.id.play_pause_toggle_view);
         if (mPlayPauseToggleImageView != null) {
             updateImageButton();
             mPlayPauseToggleImageView.setOnClickListener(new View.OnClickListener() {
