@@ -15,10 +15,11 @@
 package com.example.android.basicandroidaccessibility;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.ImageButton;
+import android.widget.Toolbar;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class ExpandTouchAreaActivity extends AppCompatActivity {
     private boolean mPlaying = false;
@@ -28,10 +29,10 @@ public class ExpandTouchAreaActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_expand_touch_area);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        setActionBar(toolbar);
 
-        mToggleImageButton = (ImageButton) findViewById(R.id.play_pause_toggle_view);
+        mToggleImageButton = findViewById(R.id.play_pause_toggle_view);
         if (mToggleImageButton != null) {
             setUI();
             mToggleImageButton.setOnClickListener(new View.OnClickListener() {
