@@ -15,14 +15,16 @@
 package com.example.android.basicandroidaccessibility;
 
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.widget.CheckBox;
 import android.widget.CompoundButton;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
+import android.widget.Toolbar;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.content.ContextCompat;
+
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 public class InsufficientContrastActivity extends AppCompatActivity {
 
@@ -35,15 +37,15 @@ public class InsufficientContrastActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_insufficient_contrast);
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        setActionBar(toolbar);
 
-        mLoremIpsumContainer = (RelativeLayout) findViewById(R.id.lorem_ipsum_container);
-        mLoremIpsumTitle = (TextView) findViewById(R.id.lorem_ipsum_title);
-        mLoremIpsumText = (TextView) findViewById(R.id.lorem_ipsum_text);
-        mAddItemFab = (FloatingActionButton) findViewById(R.id.add_item_fab);
+        mLoremIpsumContainer = findViewById(R.id.lorem_ipsum_container);
+        mLoremIpsumTitle = findViewById(R.id.lorem_ipsum_title);
+        mLoremIpsumText = findViewById(R.id.lorem_ipsum_text);
+        mAddItemFab = findViewById(R.id.add_item_fab);
 
-        CheckBox colorContrastCheckbox = (CheckBox) findViewById(R.id.color_contrast_checkbox);
+        CheckBox colorContrastCheckbox = findViewById(R.id.color_contrast_checkbox);
         if (colorContrastCheckbox != null) {
             colorContrastCheckbox.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
                 @Override
